支持通过后端打开云端播放器

+++ socialuni-server/socialuni-music-sdk/src/main/java/com/socialuni/social/music/sdk/model/RO/AgoraPlayMusicPlayerRO.java
@@ -6,6 +6,6 @@ import lombok.Data;
 public class AgoraPlayMusicPlayerRO {
     String id;
     String status;
-    Integer uid;
-    Integer createTs;
+    Long uid;
+    Long createTs;
 }
